add test cases

+++ src/main/scala/core/BreakerOptions.scala
@@ -3,7 +3,7 @@ package core
 import pureconfig._
 import pureconfig.generic.auto._
 
-case class BreakerOptions(maxBreakerFailures: Int, resetTimeoutSecs: Int, breakerDescription: String)
+case class BreakerOptions(maxBreakerFailures: Int, resetTimeoutSecs: Long, breakerDescription: String)
 
 object BreakerOptions {
   val breakerOptions = ConfigSource.default.loadOrThrow[BreakerOptions]
